lambda/requirements.txt
ask-sdk==1.13.0
ask-sdk-core==1.13.0
ask-sdk-dynamodb-persistence-adapter==1.13.0
ask-sdk-model==1.23.0
ask-sdk-runtime==1.13.0
autopep8==1.5.2
boto3==1.12.39
botocore==1.15.39
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
docutils==0.15.2
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.14.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
google-cloud-pubsub==1.4.3
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.28.1
httplib2==0.17.2
idna==2.9
jmespath==0.9.5
oauthlib==3.1.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pynamodb==4.3.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
